+++ FluidSim/Vorton.cpp
@@ -1,7 +1,7 @@
 
 #include "Vorton.h"
 
-Vorton::Vorton(const TriangleNetObject &triangleNetObject, const glm::vec3 &initialPosition, const glm::vec3 &vorticity, float radius)
+Vorton::Vorton(const TriangleNetObject &triangleNetObject, const glm::dvec3 &initialPosition, const glm::dvec3 &vorticity, double radius)
 	: TriangleNetObject(triangleNetObject), Object(triangleNetObject), m_Vorticity(vorticity), m_Radius(radius), m_Velocity(0)
 {
 	position(initialPosition);
@@ -16,33 +16,33 @@ Vorton::~Vorton()
 {
 }
 
-glm::vec3 Vorton::inducedVelocity(const glm::vec3 & querriedPosition) const
+glm::dvec3 Vorton::inducedVelocity(const glm::dvec3 & querriedPosition) const
 {
 	/*if (m_Vorticity.length < SIGNIFICANT_VORTICITY) {
 		return glm::vec3(0);
 	}*/
-	glm::vec3 distanceVector = querriedPosition - position();
-	float distanceMagnitude = static_cast<float>(glm::length(distanceVector));
-	float divisor = (distanceMagnitude < m_Radius) ? (pow(m_Radius, 2) * distanceMagnitude) : pow(distanceMagnitude, 3);
+	glm::dvec3 distanceVector = querriedPosition - position();
+	double distanceMagnitude = glm::length(distanceVector);
+	double divisor = (distanceMagnitude < m_Radius) ? (pow(m_Radius, 2) * distanceMagnitude) : pow(distanceMagnitude, 3);
 	return glm::cross(ONE_OVER_FOUR_PI * 8.0f * pow(m_Radius, 3) * m_Vorticity, distanceVector / divisor);
 }
 
-glm::vec3 Vorton::vorticity() const
+glm::dvec3 Vorton::vorticity() const
 {
 	return m_Vorticity;
 }
 
-float Vorton::radius() const
+double Vorton::radius() const
 {
 	return m_Radius;
 }
 
-void Vorton::radius(float radius)
+void Vorton::radius(double radius)
 {
 	m_Radius = radius;
 }
 
-void Vorton::vorticity(const glm::vec3 & vorticity)
+void Vorton::vorticity(const glm::dvec3 & vorticity)
 {
 	m_Vorticity = vorticity;
 }
