+++ FluidSim/VortonOctHeap.h
@@ -17,45 +17,45 @@ public:
 	VortonOctHeapElement& root();
 	std::pair<std::vector<VortonOctHeapElement>::iterator, std::vector<VortonOctHeapElement>::iterator> leafs();
 	VortonOctHeapElement& atIndex(size_t index);
-	VortonOctHeapElement& leaftAtPosition(const glm::vec3 &position);
+	VortonOctHeapElement& leaftAtPosition(const glm::dvec3 &position);
 
-	glm::vec3 minCorner() const;
-	glm::vec3 extent() const;
+	glm::dvec3 minCorner() const;
+	glm::dvec3 extent() const;
 
 private:
 	void calculateBoundingBox(const std::vector<Vorton> &vortons);
-	void subdivide(float maxVolume);
+	void subdivide(double maxVolume);
 	void subdivide(size_t divisions);
 	void createEmptyOctHeap();
 	size_t calculateNeededHeapSize();
 	void initializeLeafs(std::vector<Vorton> &vortons);
 	void initializeParents();
 
-	size_t leafIndexForPosition(const glm::vec3 &position);
+	size_t leafIndexForPosition(const glm::dvec3 &position);
 	size_t indexForIndices(const glm::uvec3 &indices);
 	glm::uvec3 indicesForIndex(size_t index);
 
 	size_t calculateFirstLeafIndex() const;
 
-	bool isInsideBoundingBox(const glm::vec3 &position);
+	bool isInsideBoundingBox(const glm::dvec3 &position);
 	
-	glm::vec3 m_MinCorner;
-	glm::vec3 m_Extent;
+	glm::dvec3 m_MinCorner;
+	glm::dvec3 m_Extent;
 	size_t m_Divisions;
 	size_t m_LeafsPerDimension;
-	glm::vec3 m_ExtentPerLeaf;
+	glm::dvec3 m_ExtentPerLeaf;
 
-	static const float m_MAX_VOLUME;
+	static const double m_MAX_VOLUME;
 	static const size_t m_DEFAULT_DIVISIONS_COUNT;
 
 	std::vector<VortonOctHeapElement> m_Heap;
 	Supervorton m_NullSupervorton;
 };
 
-inline bool VortonOctHeap::isInsideBoundingBox(const glm::vec3 & position)
+inline bool VortonOctHeap::isInsideBoundingBox(const glm::dvec3 & position)
 {
 	return (
 		(!fsmath::anyLess(position, m_MinCorner))
-		&& (!fsmath::anyLess(m_MinCorner + glm::vec3(m_Extent), position))
+		&& (!fsmath::anyLess(m_MinCorner + glm::dvec3(m_Extent), position))
 		);
 }
